fix(app): fixes the issue with view not scrolling to top when for some pages
<header>
fixes the issue with view not scrolling to top when for some
pages
</header>
<body>
This commit fixes he issue with rendering a
component after the previous page is scrolled to bottom. The new page
also stays at the bottom currently.
This commit changes that and keeps
checking for Router's NavigationEnd event and once received, scrolls the
page to top smoothly
</body>